refactor: MVTとMLTのベクタータイル処理を共通化
