+++ nusamai/src/sink/vector_tile/mod.rs
@@ -1,0 +1,14 @@
+//! Shared processing primitives for vector tile sinks.
+
+pub(crate) mod geometry;
+pub(crate) mod model;
+pub(crate) mod pipeline;
+pub(crate) mod slice;
+pub(crate) mod sort;
+pub(crate) mod tile_id;
+
+pub(crate) use pipeline::{
+    generate_stage, run_directory_pipeline, slice_stage, EncodedTile, TileEncoder,
+    TilePipelineOptions, DEFAULT_MAX_COMPRESSED_TILE_SIZE, FEATURE_CHANNEL_CAPACITY,
+};
+pub(crate) use sort::feature_sorting_stage;
